backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "mysql>=0.0.3",
    "mysql-connector-python>=9.4.0",
    "oracledb>=3.3.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "werkzeug>=3.1.3",
]
